requirements.txt
asgiref==3.11.1
Django==6.0.2
django-appconf==1.2.0
django-imagekit==6.0.0
pilkit==3.0
pillow==12.1.1
python-decouple==3.8
sqlparse==0.5.5
